Code-split carousel code
